requirements.txt
httpx[http2]>=0.28.0,<1.0.0
pyyaml>=6.0,<7.0
jinja2>=3.1,<4.0
rich>=13.0,<15.0
reportlab>=4.4,<5.0
